Refactor round_up_to_hour into common.py
