+++ app/schemas/api/asset_position.py
@@ -2,17 +2,7 @@ from datetime import datetime, timedelta
 
 from pydantic import BaseModel, ConfigDict
 
-
-def round_up_to_hour(dt: datetime) -> datetime:
-    """Helper function to truncate seconds and microseconds, and round up to the next hour."""
-    # Truncate seconds and microseconds
-    dt = dt.replace(second=0, microsecond=0)
-    # If minutes are non-zero, round up to the next hour
-    if dt.minute > 0:
-        dt += timedelta(hours=1)
-    # Reset minutes to 0
-    dt = dt.replace(minute=0)
-    return dt
+from app.schemas.api.common import round_up_to_hour
 
 
 class AssetPositionCreate(BaseModel):
